refactor: shared getSoftwareDetailItems
