+++ components/SoftwareAccordion.tsx
@@ -1,15 +1,7 @@
 import Image from "next/image"
 import { type AccordionItemProps } from "@radix-ui/react-accordion"
 
-import {
-  SoftwareItem,
-  Vendor,
-  Zkvm,
-  ZkvmMetrics,
-  ZkvmVersion,
-} from "@/lib/types"
-
-import { ZKVM_THRESHOLDS } from "@/lib/constants"
+import { Vendor, Zkvm, ZkvmMetrics, ZkvmVersion } from "@/lib/types"
 
 import {
   Accordion,
@@ -25,11 +17,7 @@ import Pizza from "./Pizza"
 import SoftwareDetails from "./SoftwareDetails"
 
 import { formatShortDate } from "@/lib/date"
-import {
-  getZkvmMetricLabel,
-  getZkvmMetricSeverityLevels,
-  getZkvmsMetrics,
-} from "@/lib/metrics"
+import { getSoftwareDetailItems, getZkvmsMetrics } from "@/lib/metrics"
 import { getSlices, getZkvmsWithUsage } from "@/lib/zkvms"
 
 const SoftwareAccordionItem = ({
@@ -45,113 +33,7 @@ const SoftwareAccordionItem = ({
   }
   metrics: ZkvmMetrics
 }) => {
-  const severityLevels = getZkvmMetricSeverityLevels(metrics)
-
-  const items: SoftwareItem[] = [
-    // Section 1 - Top charts
-    {
-      id: "verification-time",
-      label: "verification times",
-      className: "col-span-2 col-start-1 row-start-1 flex-1 pt-8 text-center",
-      popoverDetails: "TODO: Popover details",
-      severity: severityLevels.verificationTime,
-      position: 7,
-      chartInfo: {
-        bestThreshold: ZKVM_THRESHOLDS.verification_ms.yellow,
-        worstThreshold: ZKVM_THRESHOLDS.verification_ms.red,
-        unit: "ms",
-        value: Number(metrics.verification_ms),
-      },
-    },
-    {
-      id: "proof-size",
-      label: "proof size",
-      className: "col-span-2 col-start-4 row-start-1 flex-1 pt-8 text-center",
-      popoverDetails: "TODO: Popover details",
-      severity: severityLevels.proofSize,
-      position: 0,
-      chartInfo: {
-        bestThreshold: ZKVM_THRESHOLDS.size_bytes.yellow / 1024,
-        worstThreshold: ZKVM_THRESHOLDS.size_bytes.red / 1024,
-        unit: "kB",
-        value: Number(metrics.size_bytes) / 1024,
-      },
-    },
-    // Section 2 - Left
-    {
-      id: "protocol-soundness",
-      label: "protocol soundness",
-      className: "col-start-2 row-start-2 text-center",
-      popoverDetails: "TODO: Popover details",
-      severity: severityLevels.protocolSoundness,
-      position: 6,
-      value: getZkvmMetricLabel(
-        severityLevels.protocolSoundness,
-        "protocol_soundness"
-      ),
-    },
-    {
-      id: "implementation-soundness",
-      label: "implementation soundness",
-      className: "col-start-2 row-start-3 text-center",
-      popoverDetails: "TODO: Popover details",
-      severity: severityLevels.implementationSoundness,
-      position: 5,
-      value: getZkvmMetricLabel(
-        severityLevels.implementationSoundness,
-        "implementation_soundness"
-      ),
-    },
-    {
-      id: "evm-stf-bytecode",
-      label: "EVM STF bytecode",
-      className: "col-start-2 row-start-4 text-center",
-      popoverDetails: "TODO: Popover details",
-      severity: severityLevels.evmStfBytecode,
-      position: 4,
-      value: getZkvmMetricLabel(
-        severityLevels.evmStfBytecode,
-        "evm_stf_bytecode"
-      ),
-    },
-    // Section 3 - Right
-    {
-      id: "security-target",
-      label: "security target",
-      className: "col-start-4 row-start-2 text-center",
-      popoverDetails: "TODO: Popover details",
-      severity: severityLevels.securityTarget,
-      position: 1,
-      value: getZkvmMetricLabel(
-        severityLevels.securityTarget,
-        "security_target_bits"
-      ),
-    },
-    {
-      id: "quantum-security",
-      label: "quantum security",
-      className: "col-start-4 row-start-3 text-center",
-      popoverDetails: "TODO: Popover details",
-      severity: severityLevels.quantumSecurity,
-      position: 2,
-      value: getZkvmMetricLabel(
-        severityLevels.quantumSecurity,
-        "quantum_security"
-      ),
-    },
-    {
-      id: "max-bounty-amount",
-      label: "bounties",
-      className: "col-start-4 row-start-4 text-center",
-      popoverDetails: "TODO: Popover details",
-      severity: severityLevels.maxBountyAmount,
-      position: 3,
-      value: getZkvmMetricLabel(
-        severityLevels.maxBountyAmount,
-        "max_bounty_amount"
-      ),
-    },
-  ]
+  const detailItems = getSoftwareDetailItems(metrics)
 
   return (
     <AccordionItem value={value} className="col-span-5 grid grid-cols-subgrid">
@@ -198,11 +80,11 @@ const SoftwareAccordionItem = ({
         </div>
 
         <AccordionTrigger className="col-start-5 my-2 h-fit gap-2 rounded-full border-2 border-primary bg-background-highlight p-0.5 pe-2 text-primary [&>svg]:size-6">
-          <Pizza slices={getSlices(items)} disableEffects />
+          <Pizza slices={getSlices(detailItems)} disableEffects />
         </AccordionTrigger>
       </div>
       <AccordionContent className="col-span-full border-b bg-gradient-to-t from-background-active p-0">
-        <SoftwareDetails items={items} />
+        <SoftwareDetails detailItems={detailItems} />
 
         <div className="flex justify-center gap-16 p-8 pt-0">
           <ButtonLink variant="outline" href={`/zkvms/${zkvm.slug}`}>
