+++ lib/metrics.ts
@@ -5,7 +5,12 @@ import {
   getZkvmSecurityMetricsByZkvmId,
 } from "./api/metrics"
 import { ZKVM_THRESHOLDS } from "./constants"
-import { SeverityLevel, ZkvmMetrics, ZkvmThresholdMetric } from "./types"
+import {
+  SeverityLevel,
+  SoftwareDetailItem,
+  ZkvmMetrics,
+  ZkvmThresholdMetric,
+} from "./types"
 
 const protocolLabels: Record<SeverityLevel, string> = {
   red: "not fully audited",
@@ -165,4 +170,116 @@ export const getZkvmMetrics = async (zkvmId: number) => {
     ...securityMetrics,
     ...performanceMetrics,
   }
+}
+
+export const getSoftwareDetailItems = (
+  metrics: ZkvmMetrics
+): SoftwareDetailItem[] => {
+  const severityLevels = getZkvmMetricSeverityLevels(metrics)
+
+  return [
+    // Section 1 - Top charts
+    {
+      id: "verification-time",
+      label: "verification times",
+      className: "col-span-2 col-start-1 row-start-1 flex-1 pt-8 text-center",
+      popoverDetails: "TODO: Popover details",
+      severity: severityLevels.verificationTime,
+      position: 7,
+      chartInfo: {
+        bestThreshold: ZKVM_THRESHOLDS.verification_ms.yellow,
+        worstThreshold: ZKVM_THRESHOLDS.verification_ms.red,
+        unit: "ms",
+        value: Number(metrics.verification_ms),
+      },
+    },
+    {
+      id: "proof-size",
+      label: "proof size",
+      className: "col-span-2 col-start-4 row-start-1 flex-1 pt-8 text-center",
+      popoverDetails: "TODO: Popover details",
+      severity: severityLevels.proofSize,
+      position: 0,
+      chartInfo: {
+        bestThreshold: ZKVM_THRESHOLDS.size_bytes.yellow / 1024,
+        worstThreshold: ZKVM_THRESHOLDS.size_bytes.red / 1024,
+        unit: "kB",
+        value: Number(metrics.size_bytes) / 1024,
+      },
+    },
+    // Section 2 - Left
+    {
+      id: "protocol-soundness",
+      label: "protocol soundness",
+      className: "col-start-2 row-start-2 text-center",
+      popoverDetails: "TODO: Popover details",
+      severity: severityLevels.protocolSoundness,
+      position: 6,
+      value: getZkvmMetricLabel(
+        severityLevels.protocolSoundness,
+        "protocol_soundness"
+      ),
+    },
+    {
+      id: "implementation-soundness",
+      label: "implementation soundness",
+      className: "col-start-2 row-start-3 text-center",
+      popoverDetails: "TODO: Popover details",
+      severity: severityLevels.implementationSoundness,
+      position: 5,
+      value: getZkvmMetricLabel(
+        severityLevels.implementationSoundness,
+        "implementation_soundness"
+      ),
+    },
+    {
+      id: "evm-stf-bytecode",
+      label: "EVM STF bytecode",
+      className: "col-start-2 row-start-4 text-center",
+      popoverDetails: "TODO: Popover details",
+      severity: severityLevels.evmStfBytecode,
+      position: 4,
+      value: getZkvmMetricLabel(
+        severityLevels.evmStfBytecode,
+        "evm_stf_bytecode"
+      ),
+    },
+    // Section 3 - Right
+    {
+      id: "security-target",
+      label: "security target",
+      className: "col-start-4 row-start-2 text-center",
+      popoverDetails: "TODO: Popover details",
+      severity: severityLevels.securityTarget,
+      position: 1,
+      value: getZkvmMetricLabel(
+        severityLevels.securityTarget,
+        "security_target_bits"
+      ),
+    },
+    {
+      id: "quantum-security",
+      label: "quantum security",
+      className: "col-start-4 row-start-3 text-center",
+      popoverDetails: "TODO: Popover details",
+      severity: severityLevels.quantumSecurity,
+      position: 2,
+      value: getZkvmMetricLabel(
+        severityLevels.quantumSecurity,
+        "quantum_security"
+      ),
+    },
+    {
+      id: "max-bounty-amount",
+      label: "bounties",
+      className: "col-start-4 row-start-4 text-center",
+      popoverDetails: "TODO: Popover details",
+      severity: severityLevels.maxBountyAmount,
+      position: 3,
+      value: getZkvmMetricLabel(
+        severityLevels.maxBountyAmount,
+        "max_bounty_amount"
+      ),
+    },
+  ]
 }
