+++ lib/zkvms.ts
@@ -1,4 +1,4 @@
-import { Slices, SoftwareItem } from "./types"
+import { Slices, SoftwareDetailItem } from "./types"
 
 import { getActiveClusterCountByZkvmId } from "@/lib/api/clusters"
 import { getZkvm, getZkvms } from "@/lib/api/zkvms"
@@ -64,7 +64,7 @@ export const getZkvmWithUsage = async ({
   }
 }
 
-export const getSlices = (items: SoftwareItem[]) =>
+export const getSlices = (items: SoftwareDetailItem[]) =>
   items
     .sort((a, b) => a.position - b.position)
     .map((item) => ({ level: item.severity })) as Slices
